+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Remove dependence on rack-flash3
 - Add security headers
 - Add session configuration with `session` method in `App`
+- Move all `_path` method definitions to `Radical::Controller`
 
 # 1.1.0 (2021-12-06)